Add change direction to supertrend indicator
Current trend change output ("changed") of supertrend indicator is not extremely useful, since it does not contain the direction of change, which is a fundamental reason for using such an indicator. While it can be calculated from the indicator values post-hoc, I instead propose adding another output "changeDirection" which equals to 1 when the indicator detects a switch from a downtrend to an uptrend and -1 when it detects a switch from an uptrend to a downtrend. This keeps code compatibility intact and only extends the amount of data provided to the user.